requirements.txt
numpy
pandas
pyarrow
hdbscan
umap-learn
scikit-learn
plotly
tqdm
pyyaml
anthropic
